refactor: extract message-input button controller to its own file
Matches how other packages keep slot controllers, e.g.
`select/src/button-controller.js`. Renames the text helper to
`hasAccessibleText`, since `aria-hidden` text is visible and
screen-reader-only text is not.
